addded acpi

+++ include/types.h
@@ -5,7 +5,7 @@
 typedef unsigned char uchar;
 typedef unsigned short ushort;
 typedef unsigned int uint;
-typedef unsigned long ulong;
+typedef unsigned long long ulong;
 
 typedef uint pde;
